LPS-153398 Break Action form into tabs
