+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectAction/tabs/BasicInfo.tsx
@@ -1,0 +1,68 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+import ClayForm, {ClayToggle} from '@clayui/form';
+import React from 'react';
+
+import {FormError} from '../../../hooks/useForm';
+import Card from '../../Card/Card';
+import Input from '../../Form/Input';
+
+export default function BasicInfo({
+	errors,
+	handleChange,
+	readOnly,
+	setValues,
+	values,
+}: IPros) {
+	return (
+		<Card title={Liferay.Language.get('basic-info')}>
+			<Input
+				error={errors.name}
+				label={Liferay.Language.get('name')}
+				name="name"
+				onChange={handleChange}
+				required
+				value={values.name}
+			/>
+
+			<Input
+				component="textarea"
+				error={errors.description}
+				label={Liferay.Language.get('description')}
+				name="description"
+				onChange={handleChange}
+				value={values.description}
+			/>
+
+			<ClayForm.Group>
+				<ClayToggle
+					disabled={readOnly}
+					label={Liferay.Language.get('active')}
+					name="indexed"
+					onToggle={(active) => setValues({active})}
+					toggled={values.active}
+				/>
+			</ClayForm.Group>
+		</Card>
+	);
+}
+
+interface IPros {
+	errors: FormError<ObjectAction & ObjectActionParameters>;
+	handleChange: React.ChangeEventHandler<HTMLInputElement>;
+	readOnly?: boolean;
+	setValues: (values: Partial<ObjectAction>) => void;
+	values: Partial<ObjectAction>;
+}
