+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/object_definitions/add_object_action.jsp
@@ -21,7 +21,7 @@ ObjectDefinitionsActionsDisplayContext objectDefinitionsActionsDisplayContext = 
 %>
 
 <react:component
-	module="js/components/ModalAddObjectAction"
+	module="js/components/ObjectAction/AddObjectAction"
 	props='<%=
 		HashMapBuilder.<String, Object>put(
 			"apiURL", objectDefinitionsActionsDisplayContext.getAPIURL()
