+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/object_definitions/edit_object_action.jsp
@@ -23,7 +23,7 @@ ObjectAction objectAction = objectDefinitionsActionsDisplayContext.getObjectActi
 %>
 
 <react:component
-	module="js/components/EditObjectAction"
+	module="js/components/ObjectAction/EditObjectAction"
 	props='<%=
 		HashMapBuilder.<String, Object>put(
 			"objectAction", objectDefinitionsActionsDisplayContext.getObjectActionJSONObject(objectAction)
