+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ObjectAction/EditObjectAction.d.ts
@@ -12,35 +12,19 @@
  * details.
  */
 
-import React from 'react';
+/// <reference types="react" />
 
-import {CustomItem} from './Form/CustomSelect/CustomSelect';
-import ObjectActionFormBase from './ObjectActionFormBase';
-
-export default function AddObjectAction({
-	apiURL,
-	objectActionExecutors = [],
-	objectActionTriggers = [],
-}: IProps) {
-	return (
-		<ObjectActionFormBase
-			objectAction={{active: true}}
-			objectActionExecutors={objectActionExecutors}
-			objectActionTriggers={objectActionTriggers}
-			requestParams={{
-				method: 'POST',
-				url: apiURL,
-			}}
-			successMessage={Liferay.Language.get(
-				'the-object-action-was-created-successfully'
-			)}
-			title={Liferay.Language.get('new-action')}
-		/>
-	);
-}
-
+import {CustomItem} from '../Form/CustomSelect/CustomSelect';
+export default function EditObjectAction({
+	objectAction: {id, ...values},
+	objectActionExecutors,
+	objectActionTriggers,
+	readOnly,
+}: IProps): JSX.Element;
 interface IProps {
-	apiURL: string;
+	objectAction: ObjectAction;
 	objectActionExecutors: CustomItem[];
 	objectActionTriggers: CustomItem[];
+	readOnly?: boolean;
 }
+export {};
