+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ObjectAction/tabs/BasicInfo.d.ts
@@ -12,28 +12,20 @@
  * details.
  */
 
-/// <reference types="react" />
-
-import 'codemirror/mode/groovy/groovy';
-import {CustomItem} from './Form/CustomSelect/CustomSelect';
-export default function ObjectActionFormBase({
-	objectAction: initialValues,
-	objectActionExecutors,
-	objectActionTriggers,
+import React from 'react';
+import {FormError} from '../../../hooks/useForm';
+export default function BasicInfo({
+	errors,
+	handleChange,
 	readOnly,
-	requestParams: {method, url},
-	successMessage,
-}: IProps): JSX.Element;
-interface IProps {
-	objectAction: Partial<ObjectAction>;
-	objectActionExecutors: CustomItem[];
-	objectActionTriggers: CustomItem[];
+	setValues,
+	values,
+}: IPros): JSX.Element;
+interface IPros {
+	errors: FormError<ObjectAction & ObjectActionParameters>;
+	handleChange: React.ChangeEventHandler<HTMLInputElement>;
 	readOnly?: boolean;
-	requestParams: {
-		method: string;
-		url: string;
-	};
-	successMessage: string;
-	title: string;
+	setValues: (values: Partial<ObjectAction>) => void;
+	values: Partial<ObjectAction>;
 }
 export {};
